+++ packages/core/src/constants/feature-flag.js
@@ -1,0 +1,1 @@
+export const ENABLE_NEW_INFO_ARCH = false