+++ packages/react-components/src/snack-bar/index.js
@@ -1,7 +1,9 @@
 import SnackBar from './components/snack-bar'
+import useSnackBar from './hooks/use-snack-bar'
 
-export { SnackBar }
+export { SnackBar, useSnackBar }
 
 export default {
   SnackBar,
+  useSnackBar,
 }